+++ src/__tests__/e2e/connection.test.ts
@@ -1,6 +1,8 @@
 if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
 import {initDriver, destroyDriver} from "../../utils/test";
 
+if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
+
 describe('Connection', () => {
     it('Test GRPC connection', async () => {
         let driver = await initDriver({endpoint: process.env.YDB_ENDPOINT || 'grpc://localhost:2136'});