+++ src/__tests__/e2e/query-service/method-execute.ts
@@ -14,6 +14,8 @@ import ExecMode = Ydb.Query.ExecMode;
 import {RetryParameters} from "../../../retries/retryParameters";
 import {RetryStrategy} from "../../../retries/retryStrategy";
 
+if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
+
 const DATABASE = '/local';
 const ENDPOINT = process.env.YDB_ENDPOINT || 'grpc://localhost:2136';
 const TABLE_NAME = 'test_table_1'