+++ src/__tests__/e2e/query-service/query-service-client.ts
@@ -7,6 +7,8 @@ import fs from "fs";
 import {AUTO_TX} from "../../../table";
 import {QuerySession, IExecuteResult} from "../../../query";
 
+if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
+
 const DATABASE = '/local';
 const ENDPOINT = process.env.YDB_ENDPOINT || 'grpc://localhost:2136';