+++ src/__tests__/e2e/query-service/transactions.ts
@@ -11,6 +11,8 @@ import {Context} from "../../../context";
 import {RetryParameters} from "../../../retries/retryParameters";
 import {RetryStrategy} from "../../../retries/retryStrategy";
 
+if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
+
 const DATABASE = '/local';
 const ENDPOINT = process.env.YDB_ENDPOINT || 'grpc://localhost:2136';