+++ src/__tests__/e2e/table-service/bulk-upsert.test.ts
@@ -4,6 +4,8 @@ import Driver from '../../../driver';
 import {TableSession} from "../../../table";
 import {Row, initDriver, destroyDriver, createTable, fillTableWithData, TABLE} from "../../../utils/test";
 
+if (process.env.TEST_ENVIRONMENT === 'dev') require('dotenv').config();
+
 async function readTable(session: TableSession): Promise<Row[]> {
     const rows: Row[] = [];